+++ es5.5-1912/CMakeLists.txt
@@ -1,0 +1,6 @@
+cmake_minimum_required(VERSION 3.27)
+project(es5_5_1912 C)
+
+set(CMAKE_C_STANDARD 99)
+
+add_executable(es5_5_1912 student.c)
